+++ app-backend/datamodel/src/test/scala/com/azavea/rf/datamodel/implicits/Generators.scala
@@ -709,13 +709,49 @@ object Generators extends ArbitraryInstances {
       }, actionType)
     }
 
+  private def toolCreateGen: Gen[Tool.Create] =
+    for {
+      title <- nonEmptyStringGen
+      description <- nonEmptyStringGen
+      requirements <- nonEmptyStringGen
+      license <- Gen.const("BSD-3")
+      visibility <- visibilityGen
+      compatibleDataSources <- Gen.const(List.empty)
+      owner <- Gen.const(None)
+      stars <- Gen.const(9999.9f) // good tools only :sunglasses:
+      definition <- Gen.const(().asJson)
+      // not super into dealing with tags or categories in testing-land right now
+      tags <- Gen.const(Seq.empty)
+      categories <- Gen.const(Seq.empty)
+    } yield {
+      Tool.Create(title,
+                  description,
+                  requirements,
+                  license,
+                  visibility,
+                  compatibleDataSources,
+                  owner,
+                  stars,
+                  definition,
+                  tags,
+                  categories)
+    }
+
   private def toolRunCreateGen: Gen[ToolRun.Create] =
     for {
       name <- Gen.option(nonEmptyStringGen)
       visibility <- visibilityGen
       executionParameters <- Gen.const(().asJson)
       owner <- Gen.const(None)
-    } yield { ToolRun.Create(name, visibility, executionParameters, owner) }
+    } yield {
+      ToolRun.Create(name,
+                     visibility,
+                     None,
+                     None,
+                     None,
+                     executionParameters,
+                     owner)
+    }
 
   private def mapTokenCreateGen: Gen[MapToken.Create] =
     nonEmptyStringGen map { name =>
@@ -864,6 +900,9 @@ object Generators extends ArbitraryInstances {
         Gen.nonEmptyListOf[ObjectAccessControlRule](
           arbitrary[ObjectAccessControlRule])
       }
+
+    implicit def arbToolCreate: Arbitrary[Tool.Create] =
+      Arbitrary { toolCreateGen }
 
     implicit def arbToolRunCreate: Arbitrary[ToolRun.Create] =
       Arbitrary { toolRunCreateGen }