+++ app-backend/db/src/main/scala/ToolRunDao.scala
@@ -36,9 +36,12 @@ object ToolRunDao extends Dao[ToolRun] with ObjectPermissions[ToolRun] {
   val selectF = sql"""
     SELECT
       distinct(id), name, created_at, created_by, modified_at, modified_by, owner, visibility,
-      execution_parameters
+      project_id, project_layer_id, template_id, execution_parameters
     FROM
   """ ++ tableF
+
+  def unsafeGetToolRunById(toolRunId: UUID): ConnectionIO[ToolRun] =
+    query.filter(toolRunId).select
 
   def insertToolRun(newRun: ToolRun.Create,
                     user: User): ConnectionIO[ToolRun] = {
@@ -48,10 +51,11 @@ object ToolRunDao extends Dao[ToolRun] with ObjectPermissions[ToolRun] {
     sql"""
           INSERT INTO tool_runs
             (id, name, created_at, created_by, modified_at, modified_by, owner, visibility,
-             execution_parameters)
+             execution_parameters, project_id, project_layer_id, template_id)
           VALUES
             (${id}, ${newRun.name}, ${now}, ${user.id}, ${now}, ${user.id}, ${newRun.owner
-      .getOrElse(user.id)}, ${newRun.visibility}, ${newRun.executionParameters})
+      .getOrElse(user.id)}, ${newRun.visibility}, ${newRun.executionParameters},
+             ${newRun.projectId}, ${newRun.projectLayerId}, ${newRun.templateId})
        """.update.withUniqueGeneratedKeys[ToolRun](
       "id",
       "name",
@@ -61,6 +65,9 @@ object ToolRunDao extends Dao[ToolRun] with ObjectPermissions[ToolRun] {
       "modified_by",
       "owner",
       "visibility",
+      "project_id",
+      "project_layer_id",
+      "template_id",
       "execution_parameters"
     )
   }