+++ app-backend/migrations/src/main/scala/migrations/Summary.scala
@@ -158,4 +158,5 @@ object MigrationSummary {
   M157
   M158
   M160
+  M161
 }